requirements.txt
pycryptodome==3.8.2
cryptography==2.7